pyproject.toml
[tool.poetry]
name = "github-pi"
version = "0.1.0"
description = ""
authors = ["Eric Prykhodko <eprykhodko@bcdtriptech.com>"]

[tool.poetry.dependencies]
python = "^3.6"
bottle = "^0.12.16"
PyGithub = "^1.43"
requests = "^2.22"
letov = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
ipython = "^7.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
